1.8 changes
* Added a new analog touch control . Improves playability on digital games also. Lets play starwars and other roms using touch screen. Digital older one is selectable in options disabling the analog stick.

* Improves digital touch control also. keeps finger tracking and not overlap other touch controls when dragged.

* Added 8-4-2-way restrictor (selectable and saved in per game options).
A restrictor gate limits the joystick's range of motion. The most common reason to use a gate in an actual arcade setting is the retrofitting of an older machine that is not compatible with a new 8-way stick. A classic example of this is Pac-Man. The game was originally designed for a 4-way stick, and is programmed to respond only when a new input occurs. If the user is holding the stick in the down position, then suddenly makes a motion to move to the right, what often happens is that the stick first moves into the down-right diagonal, which the game does not recognize as new input since down is still being held. However, right is also now considered held, and when the user completes the motion to move right, it is also not a new input, and Pac-Man will still be moving down.

*new iPad control layout.

*changed load screens.

-> changes from Bryn Thomson

some nice new icons (wait for more skin changes on release version)

-> changes from Simon Quinn

* correct rate, stereo, fps from MAME engine are processes correctly by the AudioQueue sound functions and buffers.
* sound going out of sync fixed
* stereo support.
* multiple sound rates supported
* 44khz stereo seems to work well with most games on iPad 1.
* it seems to have fixed jerky gameplay, possibly it was the AudioUnit code causing that. Much smoother.